refactor(daemon): separate gateway status projection
